+++ packages/gamut-labs/src/experimental/index.ts
@@ -1,3 +1,6 @@
 export * from './Card';
-export * from './Text';
 export * from './Popover';
+export * from './Text';
+export * from './AppHeader/AppHeaderElements/AppHeaderLogo';
+export * from './AppHeader/AppHeaderElements/AppHeaderLink';
+export * from './AppHeader/AppHeaderElements/AppHeaderDropdown';