+++ packages/gamut/src/Anchor/index.tsx
@@ -1,5 +1,25 @@
+import { properties, variant } from '@codecademy/gamut-styles';
+import { HandlerProps } from '@codecademy/gamut-system';
 import styled from '@emotion/styled';
 
-export const Anchor = styled.a`
-  text-decoration: underline;
+const anchorVariants = variant({
+  inline: {
+    textDecoration: 'underline',
+    textColor: 'hyper',
+  },
+  interface: {
+    textDecoration: 'none',
+  },
+});
+
+export type AnchorProps = HandlerProps<typeof anchorVariants> &
+  HandlerProps<typeof properties.textColor>;
+
+export const Anchor = styled.a<AnchorProps>`
+  ${properties.textColor}
+  ${anchorVariants}
 `;
+
+Anchor.defaultProps = {
+  variant: 'inline',
+};